# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vertex AI Mailer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads/parse:
    post:
      operationId: parseUploadedFile
      tags: [uploads]
      summary: Parse CSV/XLSX file and return detected leads
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "200":
          description: Parsed leads preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedFileResult"

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FileUploadInput:
      type: object
      properties:
        filename:
          type: string
      required: [filename]
    ParsedLeadRow:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        vehicle:
          type: ["string", "null"]
        route:
          type: ["string", "null"]
        pickup:
          type: ["string", "null"]
        delivery:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        hasValidEmail:
          type: boolean
        isDuplicate:
          type: boolean
      required: [hasValidEmail, isDuplicate]
    ParsedFileResult:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ParsedLeadRow"
        totalRows:
          type: integer
        validRows:
          type: integer
        invalidRows:
          type: integer
        duplicateRows:
          type: integer
        detectedColumns:
          type: array
          items:
            type: string
      required: [rows, totalRows, validRows, invalidRows, duplicateRows, detectedColumns]

    # ─── ADMIN ───────────────────────────────────────────────────────────────
...