# lib/api-zod/src/generated/types/parsedLeadRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vertex AI Mailer API
 * OpenAPI spec version: 0.1.0
 */

export interface ParsedLeadRow {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  vehicle?: string | null;
  /** @nullable */
  route?: string | null;
  /** @nullable */
  pickup?: string | null;
  /** @nullable */
  delivery?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
  hasValidEmail: boolean;
  isDuplicate: boolean;
}